requirements.txt
fastapi
uvicorn
streamlit
chromadb
torch
transformers
ffmpeg-python
pillow
python-multipart
openai-whisper
numpy
requests

